Add personal notice to AWS Region & AZ post
A brief personal note was added to the beginning of the AWS Region & AZ blog post. This notation serves to clarify that the document is not intended to disseminate information to the general public, but rather is a place for the author to compile and readily refer to personal learnings.